QUAL-115: Extended ThalesZMQMessage to be able to store either a serialized or unserialized message body, and set up the property accessors to automatically convert if possible.
Added an optional constructor parameter to ThalesZMQMessage
which will simplify the most common use case of creating a
message to wrap a request or response GPB object, like this:

  response = ThalesZMQMessage("CPULoadingResponse", cpuLoadingResponse)

Modified ThalesZMQClient and ThalesZMQServer to print a
message if a received message has more than 3 parts, but
otherwise proceed and attempt to parse it.  This will be
converted to a log message once we have logging.